requirements/base.txt
Django==3.0.6
djangorestframework==3.11.0
django-extensions==2.2.9
django-cors-headers==3.2.1
factory-boy==2.12.0
Pillow==7.1.2
pandas==1.0.4
Cython==0.29.19
django-celery-beat==2.0.0
redis==3.5.3
django-celery-results
python-dotenv
factory-boy==2.12.0
